app/requirements.txt
matplotlib==3.6.2
numpy==1.23.2
pandas==1.4.3
Pillow==9.3.0
requests==2.28.1
streamlit==1.15.2
streamlit-lottie==0.0.3